Add themes

+++ include/Connect4/game/board.h
@@ -33,6 +33,6 @@ void C4_Board_Destroy(C4_Board* board);
 void C4_Board_Reset(C4_Board* board);
 C4_SlotState C4_Board_GetSlot(C4_Board* board, uint8_t x, uint8_t y);
 bool C4_Board_SetSlot(C4_Board* board, uint8_t x, uint8_t y, C4_SlotState state);
-int64_t C4_Board_DoMove(C4_Board* board, uint8_t inColumn);
+int C4_Board_DoMove(C4_Board* board, uint8_t inColumn);
 void C4_Board_UpdateTestStr(C4_Board* board, char* buffer, size_t bufferSize);
 C4_SlotState C4_Board_GetWinner(C4_Board* board, size_t mostRecentMoveIndex);
